stem_splitting/requirements.txt
Flask
google-cloud-storage
python-dotenv
demucs
torch
torchaudio